requirements.txt
torch>=1.13.1
transformers>=4.37.2
datasets>=2.14.3
accelerate>=0.27.2
peft>=0.10.0
trl>=0.8.1
gradio>=4.0.0
scipy
einops
sentencepiece
protobuf
uvicorn
pydantic
fastapi
sse-starlette
matplotlib
fire
zhipuai
sentence_transformers
faiss-cpu
xlrd
openai
rouge
jieba
fastchat
backoff
langchain
chromadb
langchain_core
langchain_openai
langgraph
langchainhub
pyyaml
python-dotenv